// frontend/src/gen/types/ClientOut.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { GeocodeStatus } from "./GeocodeStatus.ts";

/**
 * ClientOut
 */
export type ClientOut = {
  /**
   * @type integer
   */
  id: number;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  phone: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @type string
   */
  address: string;
  /**
   * @type string
   */
  notes: string;
  /**
   * @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
   * @type string
   */
  rate: string;
  /**
   * @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
   * @type string
   */
  cost: string;
  /**
   * @type boolean
   */
  is_active: boolean;
  lat: number | null;
  lng: number | null;
  /**
   * @type string
   */
  geocode_status: GeocodeStatus;
  /**
   * @type string, date-time
   */
  created_at: string;
  /**
   * @type string, date-time
   */
  updated_at: string;
};